requirements.txt
pika
pymongo
paho-mqtt